# poc-integracion/src/main/resources/openapi/openapi-creditcard.yaml
openapi: 3.0.3
info:
  title: Bank Credit Card API
  version: 1.0.0
  description: Contrato API-first para la gestion de tarjetas de credito.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/credit-cards:
    post:
      tags: [CreditCards]
      operationId: createCreditCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCreditCardRequest'
      responses:
        '201':
          description: Tarjeta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardEnvelope'
        '422':
          description: Error de validacion o fraude
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/credit-cards/{externalId}:
    get:
      tags: [CreditCards]
      operationId: getCreditCardByExternalId
      parameters:
        - name: externalId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tarjeta encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardEnvelope'
        '404':
          description: Tarjeta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
    CreateCreditCardRequest:
      type: object
      required: [holderName, cardNumber, documentNumber]
      properties:
        holderName:
          type: string
          minLength: 3
          maxLength: 150
        cardNumber:
          type: string
          minLength: 13
          maxLength: 19
        documentNumber:
          type: string
          minLength: 8
          maxLength: 20
    CreditCard:
      type: object
      required: [id, externalId, holderName, maskedCardNumber, maskedDocumentNumber, status]
      properties:
        id:
          type: integer
          format: int64
        externalId:
          type: string
        holderName:
          type: string
        maskedCardNumber:
          type: string
        maskedDocumentNumber:
          type: string
        status:
          type: string
          enum: [APPROVED, REJECTED]
    ApiMetadata:
      type: object
      properties:
        traceId:
          type: string
        correlationId:
          type: string
        requestId:
          type: string
        apiVersion:
          type: string
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      required: [code, message, httpStatus]
      properties:
        code:
          type: string
        message:
          type: string
        httpStatus:
          type: integer
        traceId:
          type: string
    CreditCardEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/CreditCard'
        error:
          $ref: '#/components/schemas/ApiError'
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        error:
          $ref: '#/components/schemas/ApiError'
        metadata:
          $ref: '#/components/schemas/ApiMetadata'